second file + changes to rlm
